# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Marketplace de Productos para Mascotas API",
        "description": "API RESTful para el marketplace de productos para mascotas. Permite gestionar usuarios, productos, tiendas, carritos y pedidos mediante autenticación Bearer Token (Sanctum).",
        "contact": {
            "name": "Equipo Marketplace Mascotas",
            "email": "admin@marketplace-mascotas.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://marketplace_productos_mascotas.test/api",
            "description": "API local"
        }
    ],
    "paths": {
...
        "/payments": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Registrar pago de un pedido",
                "description": "Crea un pago asociado a una orden. El monto se toma automáticamente del total de la orden.",
                "operationId": "storePayment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "method",
                                    "transaction_ref"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "method": {
                                        "type": "string",
                                        "example": "card",
                                        "enum": [
                                            "cash",
                                            "card",
                                            "transfer"
                                        ]
                                    },
                                    "transaction_ref": {
                                        "type": "string",
                                        "example": "TXN123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pago registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "method": {
                                                    "type": "string",
                                                    "example": "card"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "transaction_ref": {
                                                    "type": "string",
                                                    "example": "TXN123456"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-04-08T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para pagar esta orden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado para pagar esta orden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "La orden ya fue pagada o error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Esta orden ya ha sido pagada."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/payments/{order_id}": {
            "get": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Ver estado del pago de un pedido",
                "description": "Devuelve el pago asociado a la orden indicada. Los buyers solo pueden ver sus propios pagos.",
                "operationId": "showPayment",
                "parameters": [
                    {
                        "name": "order_id",
                        "in": "path",
                        "description": "ID de la orden",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del pago",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "order_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "method": {
                                                    "type": "string",
                                                    "example": "card"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 150
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                },
                                                "transaction_ref": {
                                                    "type": "string",
                                                    "example": "TXN123456"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-04-08T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado para ver este pago",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado para ver este pago."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay pago registrado para este pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No hay pago registrado para este pedido."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/payments/{id}/confirm": {
            "patch": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Confirmar pago manualmente",
                "description": "Permite a un administrador o vendedor confirmar un pago pendiente.",
                "operationId": "confirmPayment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pago",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pago confirmado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "confirmed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sin permisos para confirmar pagos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Pago no encontrado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "Ingresá el token obtenido en /api/login. Formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}